pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.5.15",
    "fastmcp>=0.2.0",
    "llama-parse>=0.4.0",
    "neo4j>=5.0.0",
    "numpy>=1.21.0",
    "pdfplumber>=0.11.7",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
    "sentence-transformers>=2.0.0",
    "torch>=2.0.0",
    "typing-extensions>=4.0.0",
]
